- did some code cleanup and refactoring - fixed bug in coordinate display conversion - --geometry option now handles size of application window - resizing of the application window also updates mapsize - removed X dependencies - made some buttons bigger for car mode
git-svn-id: https://gpsdrive.svn.sourceforge.net/svnroot/gpsdrive/trunk@1787 e3ca9d6f-f84a-410d-b97a-e7ddc331358d

+++ src/gpsproto.h
@@ -87,11 +87,11 @@ gint setutc (GtkWidget * widget, guint datum);
 G_MODULE_EXPORT gint modulesetup ();
 gint message_cb (char *msgid, char *name, char *text, int fs);
 void exiterr (int exitcode);
-void calcxy (gdouble * posx, gdouble * posy, gdouble lon, gdouble lat, gint zoom);
+void calcxy (gint * posx, gint * posy, gdouble lon, gdouble lat, gint zoom);
 void minimap_xy2latlon(gint x, gint y, gdouble *lon, gdouble *lat, gdouble *dif);
 void rebuildtracklist (void);
 gint error_popup (gpointer datum);
-void calcxymini (gdouble * posx, gdouble * posy, gdouble lon, gdouble lat,
+void calcxymini (gint * posx, gint * posy, gdouble lon, gdouble lat,
 	    gint zoom);
 gdouble calcR (gdouble lat);
 void calcxytopos (int , int , gdouble *, gdouble *, int );
